refactor(db) 5/13: extract db/item.rs (closes #251)
Move 6 item populators (itemization, stats, granted abilities, sets, relic
procs), the relic classifiers, item CREATE TABLEs (into item::create_tables),
and 2 item tests into db/item.rs. Drop the colliding schema::item re-export
from mod.rs; item.rs imports crate::schema::item directly. 240 tests + clippy
-D warnings + fmt green.
